+++ .prettierignore
@@ -1,0 +1,6 @@
+android
+ios
+lib
+.vscode
+node_modules
+package.json
