+++ src/components/CallingCodePicker/index.ts
@@ -1,0 +1,1 @@
+export { default as CallingCodePicker } from './CallingCodePicker';
