+++ src/components/Flag/index.ts
@@ -1,0 +1,1 @@
+export { default as Flag } from './Flag';
