+++ src/components/PickerItem/index.ts
@@ -1,0 +1,1 @@
+export { default as PickerItem } from './PickerItem';
