+++ src/components/PickerToggler/styles.ts
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  innerContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+  imageContainer: {
+    width: 40,
+    height: 30,
+  },
+});
