+++ src/components/StyledText/StyledText.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { StyleProp, StyleSheet, Text, TextStyle } from 'react-native';
+import { StyleProp, Text, TextStyle } from 'react-native';
 
-import { colors } from '../theme';
+import styles from './styles';
 
 interface IStyledTextProps {
   style?: StyleProp<TextStyle>;
@@ -12,11 +12,3 @@ const StyledText: React.FC<IStyledTextProps> = ({ children, style }) => {
 };
 
 export default StyledText;
-
-const styles = StyleSheet.create({
-  text: {
-    fontSize: 16,
-    lineHeight: 24,
-    color: colors.black,
-  },
-});
