+++ src/components/StyledText/index.ts
@@ -1,0 +1,1 @@
+export { default as StyledText } from './StyledText';
