+++ src/components/StyledText/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+import { colors } from '../../theme';
+
+export default StyleSheet.create({
+  text: {
+    fontSize: 16,
+    lineHeight: 24,
+    color: colors.black,
+  },
+});
