--- src/components/index.ts
@@ -1,5 +1,0 @@
-export { default as PickerItem } from './PickerItem';
-export { default as PickerToggler } from './PickerToggler';
-export { default as Flag } from './Flag';
-export { default as StyledText } from './StyledText';
-export { default as Search } from './Search';
